feat: refactor helper service
